ks.cfg: Better handling of multiple IWE partitions
This change simplifies the handling of multiple partitions used for
installation keys. Now partitions with the labels IWE_INST (default
partition from the installation media) and IWE (from removable storage
made by the user) are monted separatedly and if more than one key is
found, the installation aborts.

The need for two partitions is that when booting from an actual CD or
DVD the trick used to kind of emulate real partitions does not work and
the user would not be able to save keys on a CD/DVD anyways...

Signed-off-by: Peter Senna Tschudin <peter.senna@gmail.com>
Signed-off-by: Peter Senna Tschudin <peter.senna@iwe.io>